Add QueuePurgeService
